Moved the side menu to the project level
