+++ apps/webapp/app/routes/_app.orgs.$organizationSlug.projects.$projectParam/route.tsx
@@ -1,7 +1,11 @@
 import { Outlet } from "@remix-run/react";
 import { RouteErrorDisplay } from "~/components/ErrorDisplay";
-import { useOrganization } from "~/hooks/useOrganizations";
+import { MainBody } from "~/components/layout/AppLayout";
+import { SideMenu } from "~/components/navigation/SideMenu";
+import { useEnvironment } from "~/hooks/useEnvironment";
+import { useIsImpersonating, useOrganization, useOrganizations } from "~/hooks/useOrganizations";
 import { useProject } from "~/hooks/useProject";
+import { useUser } from "~/hooks/useUser";
 import { type Handle } from "~/utils/handle";
 import { v3ProjectPath } from "~/utils/pathBuilder";
 
@@ -15,9 +19,27 @@ export const handle: Handle = {
 };
 
 export default function Project() {
+  const organizations = useOrganizations();
+  const organization = useOrganization();
+  const project = useProject();
+  const environment = useEnvironment();
+  const user = useUser();
+  const isImpersonating = useIsImpersonating();
+
   return (
     <>
-      <Outlet />
+      <div className="grid grid-cols-[14rem_1fr] overflow-hidden">
+        <SideMenu
+          user={{ ...user, isImpersonating }}
+          project={project}
+          environment={environment}
+          organization={organization}
+          organizations={organizations}
+        />
+        <MainBody>
+          <Outlet />
+        </MainBody>
+      </div>
     </>
   );
 }
